moved getWeek and getYear to utils.ts
