+++ backend/repository/data.ts
@@ -1,4 +1,4 @@
-import { range, mergeEmployees, EmployeeInformation } from './util'
+import { range, mergeEmployees, EmployeeInformation, getYear, getWeek } from './util'
 import { v4 as uuid } from 'uuid'
 
 /**
@@ -41,21 +41,6 @@ const getStorageUrl = (key: string) => {
   else {
     return undefined
   }
-}
-const getYear = () => {
-  let currentDate = new Date()
-  let currentYear = currentDate.getFullYear()
-
-  return Number(currentYear);
-}
-
-const getWeek = () => { 
-  let currentDate = new Date();
-  let oneJan = new Date(getYear(), 0, 1)
-  let numberOfDays = Math.floor((Number(currentDate) - Number(oneJan)) / (24 * 60 * 60 * 1000))
-  let currentWeekNumber = Math.floor((currentDate.getDay() + 1 + numberOfDays) / 7)
-
-  return currentWeekNumber.toString();
 }
 
 const statusColorCode = (wantNewProject, openForNewProject, inProject) => {
